Changed reflection profiles to be flex.doubles rather than flex.ints. I've tried to locate any bits that need to be changed but there may still be bits I haven't found. I'll continue to check that everything still works.

+++ algorithms/integration/copy_reflection_profiles.py
@@ -61,7 +61,7 @@ def copy_image_pixels(sweep, reflections, frame_indices,
     # If gain map or dark map are None then set suitable defaults
     image_size = sweep.get_image_size()[::-1]
     if not gain_map:
-        gain_map = flex.int(flex.grid(*image_size), 1)
+        gain_map = flex.double(flex.grid(*image_size), 1)
     if not dark_map:
         dark_map = flex.int(flex.grid(*image_size), 0)
